+++ shared/chat/conversation/input-area/mention-handler-hoc/index.js.flow
@@ -1,6 +1,6 @@
 // @flow
 import * as React from 'react'
-import {type PreMentionHocProps, type Props} from '../normal'
+import {type PreMentionHocProps, type Props} from '../normal/platform-index'
 
 declare function mentionHoc(input: React.ComponentType<Props>): React.ComponentType<PreMentionHocProps>
