+++ shared/chat/conversation/input-area/normal/index-shared.js
@@ -5,7 +5,7 @@ import {Input as TextInput} from '../../../../common-adapters'
 import * as Types from '../../../../constants/types/chat2'
 import {isMobile} from '../../../../util/container'
 import mentionHoc from '../mention-handler-hoc'
-import {default as _Input, type PreMentionHocProps} from '.'
+import {default as _Input, type PreMentionHocProps} from './platform-index'
 import {throttle} from 'lodash-es'
 import {formatTextForQuoting} from '../../../../util/chat'
