+++ shared/chat/conversation/input-area/normal/platform-index.js.flow
